+++ stdlib/InteractiveUtils/test/runtests.jl
@@ -777,3 +777,16 @@ end
 @testset "Docstrings" begin
     @test isempty(Docs.undocumented_names(InteractiveUtils))
 end
+
+@testset "@repeat" begin
+    counter = 0
+    @repeat 3 counter += 1
+    @test counter == 3
+
+    counter = 0
+    @repeat counter < 5 counter += 1
+    @test counter == 5
+
+    # cannot test because we cannot interrupt the infinite loop
+    # @repeat counter += 1
+end